requirements.txt
streamlit==1.49.1
pandas==2.3.2
numpy==2.3.3
scikit-learn==1.7.2
lightgbm==4.6.0
catboost==1.2.8
xgboost==3.0.5
joblib==1.5.2
plotly==6.3.0
seaborn==0.13.2
matplotlib==3.10.6
gspread==6.1.4
google-auth==2.35.0
google-auth-oauthlib==1.2.1
langchain-google-genai==3.1.0
langchain==1.0.8
langchain-core==1.0.6